Support multiple New Relic apps
Allow the New Relic app ID to be specified as a param to the notification, so that we can use this app to notify New Relic for multiple apps. This places the configuration of the New Relic app ID within each Heroku app's deploy hook, rather than having an instance of this app for each of them.